% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example Time Series Data from Human Connectome Project}
\format{A \code{list} with 10 elements, where each element has 
4800 rows and 25 columns corresponding to 4800 time points 
and 25 regions of interest from independent components analysis (ICA).}
\usage{
dat
}
\description{
A \code{list} of 10 subjects with time series of regions
of interest of the brain
}
\keyword{datasets}

