% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prep_bhm_data_from_mixture}
\alias{.prep_bhm_data_from_mixture}
\title{Prepare Stan data from mixture}
\usage{
.prep_bhm_data_from_mixture(mixture, pri, centered)
}
\description{
Prepare Stan data from mixture
}
\keyword{internal}
