% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prep_bhm_data_from_mle}
\alias{.prep_bhm_data_from_mle}
\title{Prepare Stan data from mle and cov}
\usage{
.prep_bhm_data_from_mle(mle, var_matrix, pri, centered)
}
\description{
Prepare Stan data from mle and cov
}
\keyword{internal}
