% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_if_lt_two_groups}
\alias{.stop_if_lt_two_groups}
\title{Internal: stop if fewer than 2 groups}
\usage{
.stop_if_lt_two_groups(groups)
}
\description{
Internal: stop if fewer than 2 groups
}
\keyword{internal}
