% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2UT.R
\name{mat2UT}
\alias{mat2UT}
\title{Matrix to Upper Triangular}
\usage{
mat2UT(x)
}
\arguments{
\item{x}{A square matrix}
}
\value{
The vectorized upper triangle of x.
}
\description{
Returns the vectorized upper triangle of a square matrix
}
\examples{
\dontrun{

}
}

