% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive.R
\name{print.shrinkr_prior_contrasts}
\alias{print.shrinkr_prior_contrasts}
\title{Print method for prior pairwise contrasts}
\usage{
\method{print}{shrinkr_prior_contrasts}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{shrinkr_prior_contrasts} object.}

\item{digits}{Number of digits to display. Default 3.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for prior pairwise contrasts
}
