% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVars.R
\name{rowVars}
\alias{rowVars}
\title{Row Variances}
\usage{
rowVars(x, na.rm = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{na.rm}{If TRUE (default), \code{NA} values are stripped before
computation.  If FALSE, \code{NA} values are included in computation.}
}
\value{
A vector of numerics of length \code{nrow(x)}
}
\description{
Computes variance of each row of a matrix.
}

