% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubEstimates.R
\name{getSubEstimates}
\alias{getSubEstimates}
\title{Computes Estimates from Sub Time Series}
\usage{
getSubEstimates(Y, FUN, block.size = 5, block.space = 1)
}
\arguments{
\item{Y}{A \eqn{T\times K} matrix, where T is the number of time points in 
the time series and K is the number of variables (e.g. voxels, regions) 
observed at each time point.}

\item{FUN}{A function that can be applied to a TxK matrix to obtain an 
estimate of the quantity of interest (e.g. \code{FUN=cor} returns the KxK 
pairwise correlation matrix)}

\item{block.size}{The number of time points forming each block for estimation 
of sampling variance (see Details).  Default is 5.}

\item{block.space}{The number of time points separating consecutive blocks 
(see Details).  Default is 1.}
}
\value{
A list containing five elements, each a vector, to be input into the 
\eqn{i}th columns of \code{W_all}, \code{W_part1}, \code{W_part2}, 
\code{W_odd} and \code{W_even}, respectively.
}
\description{
This function computes 5 estimates of the quantity(ies) of 
interest from the time series of a single subject.  The resulting estimates 
should be vectorized and concatenated into 5 matrices across subjects for 
input to shrinkIt (see \code{link{shrinkIt}}).
}
\details{
The main arguments to \code{\link{shrinkIt}} are W_all, W_part1, W_part2, 
W_odd and W_even, each a matrix size \eqn{M\times n}, where M is the length 
of the vectorized quantity of interest (e.g. number of elements in the upper 
triangle of a \eqn{K\times K} correlation matrix).  W_odd and W_even are 
used to estimate sampling variance, while W_part1 and W_part2 are used to 
estimate intrasession variance. For more details, see \code{\link{shrinkIt}}.  
This function subsamples the time series of a single subject \eqn{i} to 
obtain the \eqn{i}th columns of W_all, W_part1, W_part2, W_odd and W_even,
respectively.  

Let \eqn{Y_i} be the \eqn{T\times K} timeseries of \eqn{K} variables for 
subject \eqn{i}.  The \eqn{i}th column of \code{W_all} is the estimate of 
\eqn{x_i} produced from the full time series \eqn{Y_i}.  The \eqn{i}th column 
of \code{W_part1} is the estimate of \eqn{x_i} produced from the first 
\eqn{T'} time points of \eqn{Y_i}, where \eqn{T'\leq T/2} is the length of 
the even block time series (see below).  The \eqn{i}th column of 
\code{W_part2} is the estimate of \eqn{x_i} produced from the last \eqn{T'} 
time points of \eqn{Y_i}. 

The \eqn{i}th column of \code{W_odd} is the estimate of \eqn{x_i} produced 
from the odd blocks (of length \code{block.size}) of \eqn{Y_i}. The \eqn{i}th 
column of \code{W_even} is the estimate of \eqn{x_i} produced from the even 
blocks (of length \code{block.size}) of \eqn{Y_i}.  Blocks are defined as 
\code{block.size} consecutive time points, separated by \code{block.space} 
time points.  These are used instead of individual time points to reduce 
dependence between estimates produced from the odd and even timeseries.
}

