% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVars.R
\name{rowVars}
\alias{rowVars}
\title{Row Variances}
\usage{
rowVars(x, na.rm = TRUE)
}
\arguments{
\item{x}{a matrix}
}
\value{
A vector of numerics, same length as the number of rows of the 
matrix
}
\description{
Computes variance of each row of a matrix.
}
\examples{
\dontrun{

}
}

