% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinkR-package.R
\name{shrinkR-package}
\alias{shrinkR}
\alias{shrinkR-package}
\title{Empirical Bayes Shrinkage of Estimates of Time Series Data}
\description{
Performs empirical Bayes shrinkage towards the 
group mean of subject-level estimates of some summary statistic of 
stationary or nonstationary time series data (e.g. functional connectivity 
matrices produced from resting-state fMRI data).
}

