% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shrinkR-package}
\alias{shrinkR}
\alias{shrinkR-package}
\title{Shrinkage in resting-state fMRI}
\description{
Performs shrinkage
using resting-state fMRI based on
\url{http://www.sciencedirect.com/science/article/pii/S105381191400408X}
but for for subject-level parcellation
}

