% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinkR-package.R
\name{shrinkR-package}
\alias{shrinkR}
\alias{shrinkR-package}
\title{Shrinkage in resting-state fMRI}
\description{
Performs shrinkage
using resting-state fMRI based on 
\url{http://www.sciencedirect.com/science/article/pii/S105381191400408X}
but for for subject-level parcellation
}

