% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UT2mat.R
\name{UT2mat}
\alias{UT2mat}
\title{Upper Triangular to Matrix}
\usage{
UT2mat(x, diag = 1)
}
\arguments{
\item{x}{A vector containing the upper triangular elements of a square, symmetric 
matrix.}

\item{diag}{A scalar equal to the diagonal values of the matrix to be returned.}
}
\value{
A symmetric matrix with the values of x in the upper and lower triangles 
and the value \code{diag} on the diagonals.
}
\description{
Returns the symmatric square matrix from a vector containing the upper 
triangular elements
}
\examples{
\dontrun{

}
}

